# docs/arch/integrations/openapi/companion-socket.yaml
...
openapi: 3.1.0
info:
  title: Merkle Companion Socket API
  version: 0.1.0
  description: >
    Local administrative and control-plane HTTP API exposed by the Merkle
    Vault Agent over a Unix domain socket at
    `unix:///run/merkle/companion.sock`.


    This surface is distinct from the MCP stdio JSON-RPC transport.
    It is consumed by:

    - The **CLI Adapter** (`merkle` CLI commands such as `unseal`, `seal`,
      `backup`, `restore`, `audit`).

    - The **MCP Adapter** process that connects to the daemon after each MCP
      session handshake to obtain a session lease, resolve Use Tokens, and
      dispatch Proxy Tool operations.

    - Privileged local tooling operating as the same OS user as the Vault
      Agent.


    **Authentication**: Callers are authenticated by Unix peer credentials
    (SO_PEERCRED / SCM_CREDENTIALS). The agent verifies `uid` and optionally
    `pid` + process name against the `allowed_consumers` list in the
    Namespace Policy. The `X-Companion-Peer` request header is a
    documentation placeholder only; real authentication is kernel-enforced
    at socket accept time.


    **Transport**: HTTP/1.1 over a Unix domain socket. No TLS; isolation is
    the OS permission boundary (`0600` socket owned by the running user).


    **Error model**: All error responses use `application/problem+json`
    (RFC 7807) with a `type` discriminator drawn from the `ProblemType`
    enum.
  license:
    name: Apache-2.0
...
servers:
...
    description: >
...
security:
  - companion-socket-peer: []
...
paths:
...
  /v1/backup:
    post:
      operationId: backup-recovery--trigger-backup
      summary: Trigger an on-demand Backup
      description: >
        Temporarily unavailable. The current identity model has no distinct,
        configured `age` recipient for the Master Key, and the agent refuses
        to create a misleading single-recipient backup. This endpoint returns
        501 until the two-recipient key-provisioning design is implemented.
      tags:
        - backup-recovery
      x-merkle-sensitivity: medium
      x-merkle-oob-required: false
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  description: >
                    Free-form note appended to the Audit Entry for this
                    backup. The agent uses `pre_rotate_snapshot` internally;
                    custom values are allowed for manual triggers.
      responses:
        "501":
          description: Backup capability is not safely configured.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/backup/snapshots:
    get:
      operationId: backup-recovery--list-snapshots
      summary: List available Backup snapshots
      description: >
        Returns metadata for existing Backup files in the configured target
        directory. Does not decrypt or validate the files; use
        `/v1/backup/restore-plan` for pre-restore validation.
      tags:
        - backup-recovery
      x-merkle-sensitivity: low
      x-merkle-oob-required: false
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Snapshot list returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - snapshots
                properties:
                  snapshots:
                    type: array
                    items:
                      $ref: "#/components/schemas/BackupSnapshot"
                  next_cursor:
                    type: string
        "503":
          description: Agent is sealed or backup directory is unreadable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/backup/restore-plan:
    post:
      operationId: backup-recovery--create-restore-plan
      summary: Create a Restore Plan from a Backup snapshot
      description: >
        Temporarily unavailable. Restore plans must be durably stored and
        bound to a verified, decrypted artifact before execution. The endpoint
        returns 501 rather than presenting a plan that cannot be resolved.
      tags:
        - backup-recovery
      x-merkle-sensitivity: high
      x-merkle-oob-required: false
      x-merkle-attestation: required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - snapshot_filename
                - mode
              properties:
                snapshot_filename:
                  type: string
                  description: >
                    Filename (not full path) of a Backup file in the
                    configured target directory. Example:
                    `merkle-bk-2026-05-22T10:00:00Z.merkle.age`.
                mode:
                  $ref: "#/components/schemas/RestoreMode"
                recovery_key_path:
                  type: string
                  description: >
                    Path to the `age` Recovery Key file if the Master Key
                    is unavailable (Disaster Recovery path).
      responses:
        "501":
          description: Restore-plan capability is not safely configured.
  /v1/backup/restore:
    post:
      operationId: backup-recovery--execute-restore
      summary: Execute a Restore Plan
      description: >-
        Temporarily unavailable. Execution requires a durable restore plan,
        verified artifact integrity, and an atomic multi-secret storage port.
        The endpoint returns 501 until those invariants are implemented.
      tags:
        - backup-recovery
      x-merkle-sensitivity: high
      x-merkle-oob-required: true
      x-merkle-attestation: required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - plan_id
                - operator_confirmation
              properties:
                plan_id:
                  type: string
                  description: >-
                    Opaque plan ID returned by /v1/backup/restore-plan.
                operator_confirmation:
                  type: object
                  description: >-
                    Two-flag confirmation required for destructive restore.
                    Both slash_command and oob_ack MUST be true.
                  required:
                    - slash_command
                    - oob_ack
                  properties:
                    slash_command:
                      type: boolean
                    oob_ack:
                      type: boolean
      responses:
        "501":
          description: Restore capability is not safely configured.

  # ---------------------------------------------------------------------------
  # Diagnostics (ADR-0024 gap matrix)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    companion-socket-peer:
      type: apiKey
      in: header
      name: X-Companion-Peer
      description: >-
        Documentation placeholder only. Real authentication is performed by
        the Vault Agent at socket accept time using Unix peer credentials.
        Peer credential matrix per platform:
          Linux   — SCM_CREDENTIALS / SO_PEERCRED: uid, gid, pid inspected.
          macOS   — LOCAL_PEERCRED (getsockopt): uid and pid inspected.
          Windows — GetNamedPipeClientProcessId: pid inspected; uid derived
                    from token owner SID.
        The agent validates the caller uid, pid, and process name against
        the allowed_consumers list in the Namespace Policy. The
        X-Companion-Peer header is not evaluated for authentication; it is
        documented here to satisfy OpenAPI's requirement for a named security
        scheme. Clients MUST NOT rely on this header as an authentication
        mechanism.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Maximum number of results to return. Default 50, max 500.
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
    CursorParam:
      name: cursor
      in: query
      description: >
        Opaque pagination cursor returned by a previous response's
        `next_cursor`.
      schema:
        type: string
  schemas:
...
    RestoreMode:
      type: string
      description: >-
        Strategy applied when the Backup contains Secrets that already
        exist in the vault. Source of truth:
        docs/arch/schemas/backup_recovery/restore_plan.cue #RestoreMode.
      enum:
        - overwrite
        - merge
        - newest_wins

    # -------------------------------------------------------------------------
    # Domain aggregates
    # -------------------------------------------------------------------------
...
    BackupSnapshot:
      type: object
      description: >
        Metadata for a single Backup file. The file is `age`-encrypted with
        two recipients: the Master public key and the Recovery Public Key.
      required:
        - filename
        - created_at
        - size_bytes
      properties:
        filename:
          type: string
          description: >
            Filename in the configured backup target directory.
            Pattern: `merkle-bk-<utc-iso8601>.merkle.age`.
          example: merkle-bk-2026-05-22T10:00:00Z.merkle.age
        created_at:
          type: string
          format: date-time
        size_bytes:
          type: integer
          minimum: 0
        namespace_count:
          type: integer
          description: >
            Number of Namespaces included in the snapshot. Present only
            when the agent can read the snapshot header without decrypting
            the full payload.
        secret_count:
          type: integer
          description: Total number of Secrets included in the snapshot.
        trigger:
          type: string
          description: >-
            How the backup was triggered. Informational only.
            Source of truth:
            docs/arch/schemas/backup_recovery/backup_scheduler.cue.
            change_triggered=on_change_count threshold reached;
            idle_triggered=on_idle inactivity; anacron_triggered=boot+elapsed
            max_interval; manual=operator request.
          enum:
            - manual
            - pre_rotate
            - change_triggered
            - idle_triggered
            - anacron_triggered
...
    ProblemType:
      type: string
      description: >-
        Discriminator for the ErrorEnvelope. Each snake_case REST code maps
        to a PascalCase MCP error_type in the MCP protocol surface.
        REST ↔ MCP canonical code mapping:
          unseal_required          ↔ UnsealRequired
          rate_limit_exceeded      ↔ RateLimitExceeded
          reveal_denied            ↔ RevealDenied
          handle_not_found         ↔ HandleNotFound
          namespace_not_bound      ↔ NamespaceNotBound
          oob_confirmation_required ↔ OobConfirmationRequired
          oob_confirmation_timeout ↔ OobConfirmationTimeout
          already_bound            ↔ AlreadyBound
          schema_validation_failed ↔ SchemaValidationFailed
          duplicate_name           ↔ DuplicateName
          duplicate_fingerprint    ↔ DuplicateFingerprint
          category_not_registered  ↔ CategoryNotRegistered
          category_mismatch        ↔ CategoryMismatch
          policy_tag_required      ↔ PolicyTagRequired
          operator_confirmation_required ↔ OperatorConfirmationRequired
          agent_sealed             ↔ AgentSealed
          agent_shutting_down      ↔ AgentShuttingDown
          namespace_not_found      ↔ NamespaceNotFound
          session_not_found        ↔ SessionNotFound
          restore_plan_expired     ↔ RestorePlanExpired
          restore_conflict         ↔ RestoreConflict
          backup_failed            ↔ BackupFailed
          unseal_authentication_failed ↔ UnsealAuthenticationFailed
          proxy_tool_not_supported_for_category
            ↔ ProxyToolNotSupportedForCategory
          expose_forbidden_for_sensitivity
            ↔ ExposeForbiddenForSensitivity
          argon2id_parameters_below_minimum
            ↔ Argon2idParametersBelowMinimum
          unseal_preconditions_failed
            ↔ UnsealPreconditionsFailed
          oob_signature_missing
            ↔ OobSignatureMissing
          oob_signature_invalid
            ↔ OobSignatureInvalid
          mlock_required_failed
            ↔ MlockRequiredFailed
          entropy_unseeded
            ↔ EntropyUnseeded
          vault_state_corrupted
            ↔ VaultStateCorrupted
          already_initialized
            ↔ AlreadyInitialized
          unseal_rollback_failed
            ↔ UnsealRollbackFailed
          invalid_value_format
            ↔ InvalidValueFormat
      enum:
        - already_initialized
        - unseal_required
        - rate_limit_exceeded
        - reveal_denied
        - handle_not_found
        - namespace_not_bound
        - oob_confirmation_required
        - oob_confirmation_timeout
        - already_bound
        - schema_validation_failed
        - duplicate_name
        - duplicate_fingerprint
        - category_not_registered
        - category_mismatch
        - policy_tag_required
        - operator_confirmation_required
        - agent_sealed
        - agent_shutting_down
        - namespace_not_found
        - session_not_found
        - restore_plan_expired
        - restore_conflict
        - backup_failed
        - unseal_authentication_failed
        - proxy_tool_not_supported_for_category
        - expose_forbidden_for_sensitivity
        - argon2id_parameters_below_minimum
        - unseal_preconditions_failed
        - oob_signature_missing
        - oob_signature_invalid
        - mlock_required_failed
        - entropy_unseeded
        - vault_state_corrupted
        - already_initialized
        - unseal_rollback_failed
        - invalid_value_format
    ErrorEnvelope:
      type: object
      description: >
        RFC 7807 Problem Details object. All error responses from the
        Companion Socket API use `Content-Type: application/problem+json`
        and conform to this schema.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          $ref: "#/components/schemas/ProblemType"
        title:
          type: string
          description: Short human-readable summary of the problem type.
          example: Handle not found
        status:
          type: integer
          description: HTTP status code.
          example: 404
        detail:
          type: string
          description: >
            Human-readable explanation of this specific occurrence. Should
            include actionable remediation hints.
          example: >-
            No Secret with handle vault://acme-backend/ssh/bastion-prod
            exists in namespace acme-backend. Verify the handle with
            vault.list.
        instance:
          type: string
          description: >
            URI reference identifying the specific occurrence. Typically
            the request path that produced the error.
        hint:
          type: string
          description: >
            Concise remediation suggestion suitable for display in a CLI.
        fields:
          type: array
          description: >
            Present for `schema_validation_failed`. Lists the CUE field
            constraints that were violated.
          items:
            type: object
            required:
              - field
              - constraint
            properties:
              field:
                type: string
              constraint:
                type: string
              received:
                type: string

    # -------------------------------------------------------------------------
    # Diagnostics (ADR-0024 gap matrix — GET /v1/agent/doctor)
    # -------------------------------------------------------------------------
...